Updated release notes. Removed GCC 4.6 from supported list of compilers. Repaired all the examples which weren't compiling. Fixed a thread sanitiser warning in the replaced test harness.

+++ example/enumerate_example.cpp
@@ -38,7 +38,7 @@ int main(void)
         list=enumeration.first.get();
         for(boost::afio::directory_entry &i : list.first)
         {
-            std::cout << i.name() << " type " << i.st_type() << std::endl;
+            std::cout << i.name() << " type " << static_cast<int>(i.st_type()) << std::endl;
         }
     } while(list.second);
     //]
